requirements.txt
Django==5.0.4
django-ckeditor==6.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pillow==10.3.0
python-dotenv==1.0.1
whitenoise-6.6.0
